combine packages to reduce public footprint
