+++ contribs/emissions/src/main/java/org/matsim/contrib/emissions/HbefaRoadTypeMapping.java
@@ -1,9 +1,7 @@
 package org.matsim.contrib.emissions;
 
-import com.google.inject.ProvidedBy;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.network.Network;
-import org.matsim.contrib.emissions.utils.EmissionUtils;
 
 /**
  * Created by molloyj on 01.12.2017.
