+++ contribs/emissions/src/main/java/org/matsim/contrib/emissions/HbefaVehicleCategory.java
@@ -1,9 +1,10 @@
 /* *********************************************************************** *
  * project: org.matsim.*
+ * HbefaVehicleCategory.java
  *                                                                         *
  * *********************************************************************** *
  *                                                                         *
- * copyright       : (C) 2016 by the members listed in the COPYING,        *
+ * copyright       : (C) 2009 by the members listed in the COPYING,        *
  *                   LICENSE and WARRANTY file.                            *
  * email           : info at matsim dot org                                *
  *                                                                         *
@@ -16,18 +17,10 @@
  *   See also COPYING, LICENSE and WARRANTY file                           *
  *                                                                         *
  * *********************************************************************** */
-
-
-package org.matsim.contrib.emissions.utils;
+package org.matsim.contrib.emissions;
 
 /**
- * Created by amit on 29/09/16.
+ * @author benjamin
+ *
  */
-@Deprecated // introduce EmissionsUtils.set/getXxx(...) first, and eventually move to Attributes.  kai, oct'18
-public enum EmissionSpecificationMarker {
-    
-    @Deprecated // introduce EmissionsUtils.set/getXxx(...) first, and eventually move to Attributes.  kai, oct'18
-    BEGIN_EMISSIONS ,
-    @Deprecated // introduce EmissionsUtils.set/getXxx(...) first, and eventually move to Attributes.  kai, oct'18
-    END_EMISSIONS ;
-}
+public enum HbefaVehicleCategory { PASSENGER_CAR, HEAVY_GOODS_VEHICLE, ZERO_EMISSION_VEHICLE, MOTORCYCLE; }
