+++ contribs/emissions/src/main/java/org/matsim/contrib/emissions/LinkHbefaMapping.java
@@ -1,7 +1,6 @@
 package org.matsim.contrib.emissions;
 
 import org.matsim.api.core.v01.network.Link;
-import org.matsim.contrib.emissions.utils.EmissionUtils;
 
 class LinkHbefaMapping extends HbefaRoadTypeMapping {
     @Override
