+++ contribs/emissions/src/main/java/org/matsim/contrib/emissions/package-info.java
@@ -22,7 +22,7 @@
  *  <li> Go to "CaseDefinition" > "New"</li>
  *  <li> Select the desired parameters:</li>
  *   <ul>
- *    <li> VEHICLE CATEGORIES: Currently, PC (passenger car), HGV (heavy goods vehicle), MOTORCYCLE and ZEV (zero emission vehicle) are supported (see {@link org.matsim.contrib.emissions.types.HbefaVehicleCategory HbefaVehicleCategory})</li>
+ *    <li> VEHICLE CATEGORIES: Currently, PC (passenger car), HGV (heavy goods vehicle), MOTORCYCLE and ZEV (zero emission vehicle) are supported (see {@link org.matsim.contrib.emissions.HbefaVehicleCategory HbefaVehicleCategory})</li>
  *    <li> COMPONENTS: MATSim will handle any components that are exported from the hbefa database</li>
  *    <li> YEARS: Choose the year of your scenario (only when exporting the mandatory average emission factors files)</li>
  *    <li> FLEET COMPOSITION: Choose "EF weighted with fleet composition" for the mandatory average emission factors files, and "EF per subsegment (without weighting)" for the optional detailed emission factors files</li>
@@ -42,11 +42,11 @@
  * Required files are:
  * <ul>
  * <li>roadTypeMappingFile: This file needs to map road types in your network to HBEFA 3.x road types.
- * Update (June'2018) one can now directly add HBEFA road type to link attributes using {@link org.matsim.contrib.emissions.utils.EmissionUtils#setHbefaRoadType(org.matsim.api.core.v01.network.Link, java.lang.String)}
+ * Update (June'2018) one can now directly add HBEFA road type to link attributes using {@link org.matsim.contrib.emissions.EmissionUtils#setHbefaRoadType(org.matsim.api.core.v01.network.Link, java.lang.String)}
  * or see {@link org.matsim.contrib.emissions.utils.EmissionsConfigGroup} for a detailed description.
  * 
  * <li>emissionVehicleFile: This data type is defined in the EmissionsConfigGroup,
- * see {@link org.matsim.contrib.emissions.utils.EmissionsConfigGroup}. The following information is surrounded by {@link org.matsim.contrib.emissions.utils.EmissionSpecificationMarker}. It is described as "definition of a vehicle
+ * see {@link org.matsim.contrib.emissions.utils.EmissionsConfigGroup}. The following information is surrounded by {@link org.matsim.contrib.emissions.EmissionSpecificationMarker}. It is described as "definition of a vehicle
  *  for every person (who is allowed to choose a vehicle in the simulation):
  *  <ul>
  *  <li> REQUIRED: Vehicle description must start with the respective HbefaVehicleCategory followed by ";"
@@ -77,12 +77,12 @@
  * some functions from {@link org.matsim.core.controler.Controler Controler}.
  *
  * <h3>Types</h3>
- * {@link org.matsim.contrib.emissions.types.HbefaVehicleAttributes} contains a default constructor, setting all values to average.
+ * {@link org.matsim.contrib.emissions.HbefaVehicleAttributes} contains a default constructor, setting all values to average.
  * This way a calculation of emissions for undefined vehicle types can be performed.
- * Any instance of {@link org.matsim.contrib.emissions.types.HbefaWarmEmissionFactorKey} contains a vehicle category, a warm pollutant; a road category,
+ * Any instance of {@link org.matsim.contrib.emissions.HbefaWarmEmissionFactorKey} contains a vehicle category, a warm pollutant; a road category,
  * a traffic situation and three vehicle attributes (technology, size class, em concept).
- * Any instance of {@link org.matsim.contrib.emissions.types.HbefaTrafficSituation} specifies one of 4 traffic situations.
- * Instances of {@link org.matsim.contrib.emissions.types.HbefaColdEmissionFactorKey} contain a vehicle category, a cold pollutant, a parking time range,
+ * Any instance of {@link org.matsim.contrib.emissions.HbefaTrafficSituation} specifies one of 4 traffic situations.
+ * Instances of {@link org.matsim.contrib.emissions.HbefaColdEmissionFactorKey} contain a vehicle category, a cold pollutant, a parking time range,
  * a distance, which is driven after parking and, again, vehicle attributes.
  * The cold/warm emission factor keys are mapped to the values of cold/warm emissions, the cold/warm emission factors.
  * <br> <br>
