+++ contribs/emissions/src/main/java/org/matsim/contrib/emissions/utils/EmissionWriter.java
@@ -19,12 +19,6 @@
  * *********************************************************************** */
 package org.matsim.contrib.emissions.utils;
 
-import java.io.BufferedWriter;
-import java.io.FileWriter;
-import java.util.Collection;
-import java.util.List;
-import java.util.Map;
-import java.util.SortedMap;
 import org.apache.log4j.Logger;
 import org.matsim.api.core.v01.Coord;
 import org.matsim.api.core.v01.Id;
@@ -32,6 +26,12 @@ import org.matsim.api.core.v01.population.Activity;
 import org.matsim.api.core.v01.population.Person;
 import org.matsim.api.core.v01.population.Plan;
 import org.matsim.api.core.v01.population.Population;
+
+import java.io.BufferedWriter;
+import java.io.FileWriter;
+import java.util.Collection;
+import java.util.Map;
+import java.util.SortedMap;
 
 /**
  * @author benjamin
