+++ contribs/emissions/src/test/java/org/matsim/contrib/emissions/TestHbefaWarmEmissionFactorKey.java
@@ -18,11 +18,14 @@
  *                                                                         *
  * *********************************************************************** */
 
-package org.matsim.contrib.emissions.types;
+package org.matsim.contrib.emissions;
 
 import org.junit.Assert;
 import org.junit.Test;
-
+import org.matsim.contrib.emissions.HbefaTrafficSituation;
+import org.matsim.contrib.emissions.HbefaVehicleAttributes;
+import org.matsim.contrib.emissions.HbefaVehicleCategory;
+import org.matsim.contrib.emissions.HbefaWarmEmissionFactorKey;
 
 
 public class TestHbefaWarmEmissionFactorKey{
